+++ generatorapi/src/main/java/com/kneelawk/kfractal/generator/api/ir/FunctionDefinition.java
@@ -4,6 +4,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
+import com.kneelawk.kfractal.generator.api.ir.instruction.IInstruction;
 import com.kneelawk.kfractal.util.KFractalToStringStyle;
 import org.apache.commons.lang3.builder.ToStringBuilder;
