--- generatorapi/src/main/java/com/kneelawk/kfractal/generator/api/ir/IInstructionVisitor.java
@@ -1,4 +1,0 @@
-package com.kneelawk.kfractal.generator.api.ir;
-
-public interface IInstructionVisitor {
-}
