+++ generatorapi/src/main/java/com/kneelawk/kfractal/generator/api/ir/instruction/IInstruction.java
@@ -1,4 +1,4 @@
-package com.kneelawk.kfractal.generator.api.ir;
+package com.kneelawk.kfractal.generator.api.ir.instruction;
 
 public interface IInstruction {
 	void accept(IInstructionVisitor visitor);
