+++ generatorapi/src/main/java/com/kneelawk/kfractal/generator/api/ir/instruction/IInstructionVisitor.java
@@ -1,0 +1,4 @@
+package com.kneelawk.kfractal.generator.api.ir.instruction;
+
+public interface IInstructionVisitor {
+}
