+++ generatorapi/src/main/java/com/kneelawk/kfractal/generator/api/ir/instruction/io/VariableReference.java
@@ -1,4 +1,4 @@
-package com.kneelawk.kfractal.generator.api.ir;
+package com.kneelawk.kfractal.generator.api.ir.instruction.io;
 
 import com.kneelawk.kfractal.util.KFractalToStringStyle;
 import org.apache.commons.lang3.builder.ToStringBuilder;
