+++ generatorapi/src/main/java/com/kneelawk/kfractal/generator/api/language/BinaryOperationStatement.java
@@ -1,7 +1,7 @@
 package com.kneelawk.kfractal.generator.api.language;
 
-import com.kneelawk.kfractal.generator.api.ir.IInstruction;
-import com.kneelawk.kfractal.generator.api.ir.IInstructionVisitor;
+import com.kneelawk.kfractal.generator.api.ir.instruction.IInstruction;
+import com.kneelawk.kfractal.generator.api.ir.instruction.IInstructionVisitor;
 
 public class BinaryOperationStatement implements IInstruction {
 	private IInstruction left;
