+++ generatorapi/src/main/java/com/kneelawk/kfractal/generator/api/language/Program.java
@@ -1,6 +1,6 @@
 package com.kneelawk.kfractal.generator.api.language;
 
-import com.kneelawk.kfractal.generator.api.ir.IInstruction;
+import com.kneelawk.kfractal.generator.api.ir.instruction.IInstruction;
 
 public class Program {
 	private IInstruction statement;
